Complete checkstyle stuff

+++ src/main/java/com/davisellwood/website/blog/BlogController.java
@@ -1,5 +1,11 @@
 package com.davisellwood.website.blog;
 
+import com.davisellwood.website.dagger.interfaces.Database;
+import com.davisellwood.website.dagger.interfaces.ObjectStore;
+import com.davisellwood.website.dagger.spring.bindings.SpringStorageProvider;
+import com.google.protobuf.InvalidProtocolBufferException;
+import com.google.protobuf.Timestamp;
+import jakarta.inject.Inject;
 import java.nio.charset.StandardCharsets;
 import java.time.Instant;
 import java.time.ZoneId;
@@ -8,9 +14,7 @@ import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-
-import javax.inject.Inject;
-
+import lombok.extern.slf4j.Slf4j;
 import org.commonmark.Extension;
 import org.commonmark.ext.footnotes.FootnoteDefinition;
 import org.commonmark.ext.footnotes.FootnotesExtension;
@@ -42,31 +46,27 @@ import org.springframework.security.authentication.AuthenticationCredentialsNotF
 import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-
-import com.davisellwood.website.dagger.interfaces.Database;
-import com.davisellwood.website.dagger.interfaces.ObjectStore;
-import com.davisellwood.website.dagger.spring.bindings.SpringStorageProvider;
-import com.google.protobuf.InvalidProtocolBufferException;
-import com.google.protobuf.Timestamp;
-
-import lombok.extern.slf4j.Slf4j;
-import proto.davisellwood.website.cheapskate.CheapSkateDatabase.Database.DBEntry;
-import proto.davisellwood.website.models.BlogPostOuterClass.BlogPost;
-import proto.davisellwood.website.models.BlogPostOuterClass.BlogPostMap;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import proto.davisellwood.website.cheapskate.CheapSkateDatabase.Database.DBEntry;
+import proto.davisellwood.website.models.BlogPostOuterClass.BlogPost;
+import proto.davisellwood.website.models.BlogPostOuterClass.BlogPostMap;
+
 
 @Slf4j
 @Controller
 public class BlogController {
     public static final String PATH_PREFIX = "blog";
 
+    private static final String BLOG_POSTS_KEY = "blog-posts";
+    private static final String BLOG_POSTS_FOLDER = "blog-posts/";
+
     private final ObjectStore objectStore;
     private final Database database;
-    
+
     @Inject
     public BlogController(SpringStorageProvider storageProvider) {
         this.objectStore = storageProvider.objectStore();
@@ -75,21 +75,20 @@ public class BlogController {
 
     @PostMapping(PATH_PREFIX + "/new")
     public ResponseEntity<String> postNewBlogImage(
-        Model model,
-        @RequestBody String body,
-        @RequestHeader("api_key") String apiKey,
-        @RequestHeader("is_public") String isPublicString,
-        @RequestHeader("blog_id") String blogId,
-        @RequestHeader("blog_title") String blogTitle
-    ) {
+            Model model,
+            @RequestBody String body,
+            @RequestHeader("api_key") String apiKey,
+            @RequestHeader("is_public") String isPublicString,
+            @RequestHeader("blog_id") String blogId,
+            @RequestHeader("blog_title") String blogTitle) {
         if (!BCrypt.checkpw(apiKey, database.get("blog-post-api-key").getStringValue())) {
             throw new AuthenticationCredentialsNotFoundException("API Key was not valid");
         }
 
-        DBEntry blogs = database.get("blog-posts");
+        DBEntry blogs = database.get(BLOG_POSTS_KEY);
         if (blogs == null) {
             blogs = DBEntry.newBuilder().setByteValue(BlogPostMap.newBuilder().build().toByteString()).build();
-            database.put("blog-posts", blogs);
+            database.put(BLOG_POSTS_KEY, blogs);
         }
 
         try {
@@ -98,26 +97,30 @@ public class BlogController {
 
             boolean isPublic = Boolean.parseBoolean(isPublicString);
             Instant currentInstant = Instant.now();
-            Timestamp currentTime = Timestamp.newBuilder().setSeconds(currentInstant.getEpochSecond()).setNanos(currentInstant.getNano()).build();
+            Timestamp currentTime = Timestamp.newBuilder().setSeconds(currentInstant.getEpochSecond())
+                    .setNanos(currentInstant.getNano()).build();
             Timestamp publishDate = currentTime;
 
             if (blogMap.containsPosts(blogId)) {
                 log.info("Updating blog with post id {}", blogId);
-                
+
                 if (updatedBlogs.get(blogId).getIsPublic()) {
                     publishDate = updatedBlogs.get(blogId).getPublishDate();
                 }
             }
 
-            if (objectStore.put("blog-posts/" + blogId, body.getBytes(StandardCharsets.UTF_16LE))) {
-                BlogPost newOrUpdatedPost = BlogPost.newBuilder().setBlogId(blogId).setTitle(blogTitle).setIsPublic(isPublic).setPublishDate(publishDate).setLastEdited(currentTime).setContentFilePath("blog-posts/" + blogId).build();
-    
-                database.put("blog-posts", DBEntry.newBuilder().setByteValue(BlogPostMap.newBuilder().putAllPosts(updatedBlogs).putPosts(blogId, newOrUpdatedPost).build().toByteString()).build());
+            if (objectStore.put(BLOG_POSTS_FOLDER + blogId, body.getBytes(StandardCharsets.UTF_16LE))) {
+                BlogPost newOrUpdatedPost = BlogPost.newBuilder().setBlogId(blogId).setTitle(blogTitle)
+                        .setIsPublic(isPublic).setPublishDate(publishDate).setLastEdited(currentTime)
+                        .setContentFilePath(BLOG_POSTS_FOLDER + blogId).build();
+
+                database.put(BLOG_POSTS_KEY, DBEntry.newBuilder().setByteValue(BlogPostMap.newBuilder()
+                        .putAllPosts(updatedBlogs).putPosts(blogId, newOrUpdatedPost).build().toByteString()).build());
                 return ResponseEntity.ok().body("Successfully uploaded blog post: " + blogTitle);
             } else {
                 return ResponseEntity.internalServerError().build();
             }
-        } catch  (Exception e) {
+        } catch (Exception e) {
             log.error("Error trying to upload new blog", e);
             return ResponseEntity.badRequest().build();
         }
@@ -125,26 +128,27 @@ public class BlogController {
 
     @PostMapping(PATH_PREFIX + "/attachFile")
     public ResponseEntity<String> postFile(
-        Model model,
-        @RequestBody byte[] body,
-        @RequestHeader("api_key") String apiKey,
-        @RequestHeader("blog_id") String blogId,
-        @RequestHeader("image_name") String imageName
-    ) {
+            Model model,
+            @RequestBody byte[] body,
+            @RequestHeader("api_key") String apiKey,
+            @RequestHeader("blog_id") String blogId,
+            @RequestHeader("image_name") String imageName) {
         log.info("ATTACHING FILE");
 
         if (!BCrypt.checkpw(apiKey, database.get("blog-post-api-key").getStringValue())) {
             throw new AuthenticationCredentialsNotFoundException("API Key was not valid");
         }
 
-        return objectStore.putPublic("blog-posts/"+blogId+"/"+imageName, body) ? ResponseEntity.ok().body("Successfully uploaded image " + imageName + " for blog " + blogId) : ResponseEntity.badRequest().build();
+        return objectStore.putPublic(BLOG_POSTS_FOLDER + blogId + "/" + imageName, body)
+                ? ResponseEntity.ok().body("Successfully uploaded image " + imageName + " for blog " + blogId)
+                : ResponseEntity.badRequest().build();
     }
-    
+
     @GetMapping(PATH_PREFIX + "/view/{blog_id}")
     public String getNewBlog(Model model, @PathVariable("blog_id") String blogId) {
         BlogPostMap blogPosts;
         try {
-            blogPosts = BlogPostMap.parseFrom(database.get("blog-posts").getByteValue());
+            blogPosts = BlogPostMap.parseFrom(database.get(BLOG_POSTS_KEY).getByteValue());
         } catch (InvalidProtocolBufferException e) {
             return "error";
         }
@@ -152,14 +156,26 @@ public class BlogController {
         BlogPost post = blogPosts.getPostsMap().get(blogId);
 
         // TODO: Inject the markdown generator and provide some methods
-        List<Extension> extensions = List.of(TablesExtension.create(), HeadingAnchorExtension.create(), ImageAttributesExtension.create(), StrikethroughExtension.create(), FootnotesExtension.create());
+        List<Extension> extensions = List.of(TablesExtension.create(), HeadingAnchorExtension.create(),
+                ImageAttributesExtension.create(), StrikethroughExtension.create(), FootnotesExtension.create());
 
         Parser parser = Parser.builder().linkProcessor(new CustomLinkProcessor(blogId)).extensions(extensions).build();
-        Node document = parser.parse(new String(objectStore.get(post.getContentFilePath()).get(), StandardCharsets.UTF_16LE));
-        HtmlRenderer renderer = HtmlRenderer.builder().attributeProviderFactory(new CustomImageAttributeProvider.CustomImageAttributeProviderFactory()).nodeRendererFactory(new CustomNodeRenderer.Factory()).extensions(extensions).build();
+        Node document = parser
+                .parse(new String(objectStore.get(post.getContentFilePath()).get(), StandardCharsets.UTF_16LE));
+        HtmlRenderer renderer = HtmlRenderer.builder()
+                .attributeProviderFactory(new CustomImageAttributeProvider.CustomImageAttributeProviderFactory())
+                .nodeRendererFactory(new CustomNodeRenderer.Factory()).extensions(extensions).build();
 
-        model.addAttribute("publishDate", ZonedDateTime.ofInstant(Instant.ofEpochSecond(post.getPublishDate().getSeconds()), ZoneId.of("America/Toronto")).format(DateTimeFormatter.ofPattern("dd MMM YYYY h:mm a")));
-        model.addAttribute("lastUpdatedDate", ZonedDateTime.ofInstant(Instant.ofEpochSecond(post.getLastEdited().getSeconds()), ZoneId.of("America/Toronto")).format(DateTimeFormatter.ofPattern("dd MMM YYYY h:mm a")));
+        model.addAttribute("publishDate",
+                ZonedDateTime
+                        .ofInstant(Instant.ofEpochSecond(post.getPublishDate().getSeconds()),
+                                ZoneId.of("America/Toronto"))
+                        .format(DateTimeFormatter.ofPattern("dd MMM yyyy h:mm a")));
+        model.addAttribute("lastUpdatedDate",
+                ZonedDateTime
+                        .ofInstant(Instant.ofEpochSecond(post.getLastEdited().getSeconds()),
+                                ZoneId.of("America/Toronto"))
+                        .format(DateTimeFormatter.ofPattern("dd MMM yyyy h:mm a")));
         model.addAttribute("title", post.getTitle());
         model.addAttribute("content", renderer.render(document));
 
@@ -183,30 +199,30 @@ public class BlogController {
         public void render(Node node) {
             Text textNode = (Text) node;
 
-            if (node.getNext() instanceof Link && node.getNext().getNext() instanceof Text) {
-                Link nextNode = (Link) node.getNext();
-                Text nextNextNode = (Text) node.getNext().getNext();
-                if (textNode.getLiteral().endsWith("[") && nextNextNode.getLiteral().startsWith("]")) {
-                    html.text(textNode.getLiteral().substring(0, textNode.getLiteral().length() - 1));
-                    html.tag("a href=" + nextNode.getDestination());
-                    html.text(nextNode.getTitle());
-                    html.tag("/a");
-                    html.text(nextNextNode.getLiteral().substring(1));
-                    node.getNext().unlink();
-                    node.getNext().unlink();
-                    return;
-                }
+            if (node.getNext() instanceof Link nextNode
+                    && node.getNext().getNext() instanceof Text nextNextNode
+                    && textNode.getLiteral().endsWith("[")
+                    && nextNextNode.getLiteral().startsWith("]")) {
+
+                html.text(textNode.getLiteral().substring(0, textNode.getLiteral().length() - 1));
+                html.tag("a href=" + nextNode.getDestination());
+                html.text(nextNode.getTitle());
+                html.tag("/a");
+                html.text(nextNextNode.getLiteral().substring(1));
+                node.getNext().unlink();
+                node.getNext().unlink();
+                return;
             }
 
             html.text(textNode.getLiteral());
         }
-        
+
         public static class Factory implements HtmlNodeRendererFactory {
             @Override
             public NodeRenderer create(HtmlNodeRendererContext context) {
                 return new CustomNodeRenderer(context);
             }
-            
+
         }
     }
 
@@ -217,7 +233,7 @@ public class BlogController {
             public AttributeProvider create(AttributeProviderContext context) {
                 return new CustomImageAttributeProvider();
             }
-            
+
         }
 
         @Override
@@ -228,7 +244,7 @@ public class BlogController {
             } else if (tagName.equals("th")) {
                 attributes.put("scope", "col");
                 if (attributes.get("align") != null) {
-                    attributes.put("style", "text-align:"+attributes.get("align")+";");
+                    attributes.put("style", "text-align:" + attributes.get("align") + ";");
                 }
                 return;
             } else if (tagName.equals("img")) {
@@ -257,9 +273,15 @@ public class BlogController {
 
         @Override
         public LinkResult process(LinkInfo linkInfo, Scanner scanner, InlineParserContext context) {
-            if (linkInfo.marker() != null && "!".equals(linkInfo.marker().getLiteral()) && linkInfo.text().startsWith("[") && linkInfo.text().endsWith("]")) {
+            if (linkInfo.marker() != null && "!".equals(linkInfo.marker().getLiteral())
+                    && linkInfo.text().startsWith("[") && linkInfo.text().endsWith("]")) {
                 // TODO: Rework this and store image locations in blog entry in db
-                return LinkResult.replaceWith(new Image("https://davisellwood-website.nyc3.cdn.digitaloceanspaces.com/blog-posts/"+blogId+"/"+linkInfo.text().substring(1, linkInfo.text().length() - 1).toString(), "title"), scanner.position()).includeMarker();
+                return LinkResult
+                        .replaceWith(new Image(
+                                "https://davisellwood-website.nyc3.cdn.digitaloceanspaces.com/blog-posts/" + blogId
+                                        + "/" + linkInfo.text().substring(1, linkInfo.text().length() - 1),
+                                "title"), scanner.position())
+                        .includeMarker();
             }
             if (linkInfo.marker() == null && linkInfo.text().startsWith("#")) {
                 String headingText = linkInfo.text().substring(1);
